fix cyclic import and move distance_law_from_mat back to hicstuff.py
