+++ hicstuff/hicstuff.py
@@ -34,9 +34,63 @@ import random
 import multiprocessing as mp
 import pandas as pd
 import sys
-import hicstuff.distance_law as hcdl
 from hicstuff.log import logger
 
+
+def distance_law_from_mat(matrix, indices=None, log_bins=True, base=1.1):
+    """Compute distance law as a function of the genomic coordinate aka P(s).
+    Bin length increases exponentially with distance if log_bins is True. Works
+    on dense and sparse matrices. Less precise than the one from the pairs.
+    Parameters
+    ----------
+    matrix : numpy.array or scipy.sparse.coo_matrix
+        Hi-C contact map of the chromosome on which the distance law is
+        calculated.
+    indices : None or numpy array
+        List of indices on which to compute the distance law. For example
+        compartments or expressed genes.
+    log_bins : bool
+        Whether the distance law should be computed on exponentially larger
+        bins.
+    Returns
+    -------
+    numpy array of floats :
+        The start index of each bin.
+    numpy array of floats :
+        The distance law computed per bin on the diagonal
+    """
+
+    n = min(matrix.shape)
+    included_bins = np.zeros(n, dtype=bool)
+    if indices is None:
+        included_bins[:] = True
+    else:
+        included_bins[indices] = True
+    D = np.array(
+        [
+            np.average(matrix.diagonal(j)[included_bins[: n - j]])
+            for j in range(n)
+        ]
+    )
+    if not log_bins:
+        return np.array(range(len(D))), D
+    else:
+        n_bins = int(np.log(n) / np.log(base) + 1)
+        logbin = np.unique(
+            np.logspace(0, n_bins - 1, num=n_bins, base=base, dtype=np.int)
+        )
+        logbin = np.insert(logbin, 0, 0)
+        logbin[-1] = min(n, logbin[-1])
+        if n < logbin.shape[0]:
+            print("Not enough bins. Increase logarithm base.")
+            return np.array(range(len(D))), D
+        logD = np.array(
+            [
+                np.average(D[logbin[i - 1] : logbin[i]])
+                for i in range(1, len(logbin))
+            ]
+        )
+        return logbin[:-1], logD
 
 
 def despeckle_simple(B, th2=2, threads=1):
@@ -1743,7 +1797,7 @@ def compartments(M, normalize=True):
         N = np.copy(M)
     # Computation of genomic distance law matrice:
     dist_mat = np.zeros((n, n))
-    _, dist_vals = hcdl.distance_law_from_mat(N, log_bins=False)
+    _, dist_vals = distance_law_from_mat(N, log_bins=False)
     for i in range(n):
         for j in range(n):
             dist_mat[i, j] = dist_vals[abs(j - i)]
@@ -1822,7 +1876,7 @@ def compartments_sparse(M, normalize=True):
         N = copy.copy(M)
     N = N.tocoo()
     # Detrend by the distance law
-    dist_bins, dist_vals = hcdl.distance_law_from_mat(N, log_bins=False)
+    dist_bins, dist_vals = distance_law_from_mat(N, log_bins=False)
     N.data /= dist_vals[abs(N.row - N.col)]
     N = N.tocsr()
     # Make matrix symmetric (in case of upper triangle)
