+++ tests/test_hicstuff.py
@@ -211,3 +211,14 @@ def test_compartments_sparse(matrix_size):
     pc1_s, pc2_s = hcs.compartments_sparse(M_s, normalize=False)
     assert np.isclose(np.abs(pc1_d), np.abs(pc1_s), rtol=0.01).all()
     assert np.isclose(np.abs(pc2_d), np.abs(pc2_s), rtol=0.01).all()
+
+
+@pytest.mark.parametrize(*SIZE_PARAMETERS)
+def test_distance_law(matrix_size):
+    M_d, M_s = _gen_matrices(matrix_size)
+
+    I_d, D_d = hcs.distance_law_from_mat(M_d)
+    I_s, D_s = hcs.distance_law_from_mat(M_s)
+
+    assert I_d.shape == I_s.shape
+    assert D_d.shape == D_s.shape
